relocate scorecard custom test example, add custom scorecard test image to Makefile
