+++ changelog/fragments/scorecard-custom-image-relocate.yaml
@@ -1,0 +1,14 @@
+entries:
+  - description: Updated custom scorecard test example
+    kind: change
+    breaking: false
+    migration:
+      header: Update custom scorecard test example
+      body: |
+        The custom-scorecard-tests image is built like the other
+        scorecard images with this change, the Makefile is updated
+        to produce the custom-scorecard-tests image.  This makes it
+        easier to build the image using the normal build process and
+        will make testing the custom image easier, which will be
+        addressed in another PR.
+
