--- internal/scorecard/examples/custom-scorecard-tests/bundle/manifests/memcached-operator.clusterserviceversion.yaml
@@ -1,179 +1,0 @@
-apiVersion: operators.coreos.com/v1alpha1
-kind: ClusterServiceVersion
-metadata:
-  annotations:
-    alm-examples: |-
-      [
-        {
-          "apiVersion": "cache.example.com/v1alpha1",
-          "kind": "Memcached",
-          "metadata": {
-            "name": "example-memcached"
-          },
-          "spec": {
-            "size": 3
-          },
-          "status": {
-            "nodes": null
-          }          
-        }
-      ]
-    capabilities: Basic Install
-  name: memcached-operator.v0.0.1
-  namespace: placeholder
-spec:
-  apiservicedefinitions: {}
-  customresourcedefinitions:
-    owned:
-    - description: Memcached is the Schema for the memcacheds API
-      kind: Memcached
-      name: memcacheds.cache.example.com
-      version: v1alpha1
-      specDescriptors:
-      - description: The desired number of memcached instances to deploy.
-        displayName: Size
-        path: size
-        x-descriptors:
-        - 'urn:alm:descriptor:com.tectonic.ui:podCount'
-      resources:
-      - kind: Pod
-        version: v1
-      statusDescriptors:
-      - description: Nodes are the names of the memcached pods
-        displayName: Nodes
-        path: nodes
-        x-descriptors:
-        - 'urn:alm:descriptor:com.tectonic.ui:nodes'           
-  description: Placeholder description
-  displayName: Memcached Operator
-  icon:
-  - base64data: ""
-    mediatype: ""
-  install:
-    spec:
-      deployments:
-      - name: memcached-operator
-        spec:
-          replicas: 1
-          selector:
-            matchLabels:
-              name: memcached-operator
-          strategy: {}
-          template:
-            metadata:
-              labels:
-                name: memcached-operator
-            spec:
-              containers:
-              - command:
-                - memcached-operator
-                env:
-                - name: WATCH_NAMESPACE
-                  valueFrom:
-                    fieldRef:
-                      fieldPath: metadata.annotations['olm.targetNamespaces']
-                - name: POD_NAME
-                  valueFrom:
-                    fieldRef:
-                      fieldPath: metadata.name
-                - name: OPERATOR_NAME
-                  value: memcached-operator
-                image: REPLACE_IMAGE
-                imagePullPolicy: Always
-                name: memcached-operator
-                resources: {}
-              serviceAccountName: memcached-operator
-      permissions:
-      - rules:
-        - apiGroups:
-          - ""
-          resources:
-          - pods
-          - services
-          - services/finalizers
-          - endpoints
-          - persistentvolumeclaims
-          - events
-          - configmaps
-          - secrets
-          verbs:
-          - create
-          - delete
-          - get
-          - list
-          - patch
-          - update
-          - watch
-        - apiGroups:
-          - apps
-          resources:
-          - deployments
-          - daemonsets
-          - replicasets
-          - statefulsets
-          verbs:
-          - create
-          - delete
-          - get
-          - list
-          - patch
-          - update
-          - watch
-        - apiGroups:
-          - monitoring.coreos.com
-          resources:
-          - servicemonitors
-          verbs:
-          - get
-          - create
-        - apiGroups:
-          - apps
-          resourceNames:
-          - memcached-operator
-          resources:
-          - deployments/finalizers
-          verbs:
-          - update
-        - apiGroups:
-          - ""
-          resources:
-          - pods
-          verbs:
-          - get
-        - apiGroups:
-          - apps
-          resources:
-          - replicasets
-          - deployments
-          verbs:
-          - get
-        - apiGroups:
-          - cache.example.com
-          resources:
-          - '*'
-          verbs:
-          - create
-          - delete
-          - get
-          - list
-          - patch
-          - update
-          - watch
-        serviceAccountName: memcached-operator
-    strategy: deployment
-  installModes:
-  - supported: true
-    type: OwnNamespace
-  - supported: true
-    type: SingleNamespace
-  - supported: false
-    type: MultiNamespace
-  - supported: true
-    type: AllNamespaces
-  keywords:
-  - ""
-  maintainers:
-  - {}
-  maturity: alpha
-  provider: {}
-  version: 0.0.1
