--- internal/scorecard/examples/custom-scorecard-tests/bundle/metadata/annotations.yaml
@@ -1,9 +1,0 @@
-annotations:
-  operators.operatorframework.io.bundle.mediatype.v1: registry+v1
-  operators.operatorframework.io.bundle.manifests.v1: manifests/
-  operators.operatorframework.io.bundle.metadata.v1: metadata/
-  operators.operatorframework.io.bundle.package.v1: memcached-operator
-  operators.operatorframework.io.bundle.channels.v1: alpha,stable
-  operators.operatorframework.io.bundle.channel.default.v1: stable
-  operators.operatorframework.io.test.mediatype.v1: scorecard+v1
-  operators.operatorframework.io.test.config.v1: tests/scorecard/
