--- internal/scorecard/examples/custom-scorecard-tests/bundle/tests/scorecard/config.yaml
@@ -1,20 +1,0 @@
-kind: Configuration
-apiversion: scorecard.operatorframework.io/v1alpha3
-metadata:
-  name: config
-stages:
-- tests:
-  - image: quay.io/username/custom-scorecard-tests:dev
-    entrypoint:
-    - custom-scorecard-tests
-    - customtest1
-    labels:
-      suite: custom
-      test: customtest1
-  - image: quay.io/username/custom-scorecard-tests:dev
-    entrypoint:
-    - custom-scorecard-tests
-    - customtest2
-    labels:
-      suite: custom
-      test: customtest2
