--- internal/scorecard/examples/custom-scorecard-tests/internal/tests/tests.go
@@ -1,53 +1,0 @@
-// Copyright 2020 The Operator-SDK Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package tests
-
-import (
-	apimanifests "github.com/operator-framework/api/pkg/manifests"
-	scapiv1alpha3 "github.com/operator-framework/api/pkg/apis/scorecard/v1alpha3"
-)
-
-const (
-	CustomTest1Name = "customtest1"
-	CustomTest2Name = "customtest2"
-)
-
-// Define any operator specific custom tests here.
-// CustomTest1 and CustomTest2 are example test functions. Relevant operator specific
-// test logic is to be implemented in similarly.
-
-func CustomTest1(bundle *apimanifests.Bundle) scapiv1alpha3.TestStatus {
-	r := scapiv1alpha3.TestResult{}
-	r.Name = CustomTest1Name
-	r.State = scapiv1alpha3.PassState
-	r.Errors = make([]string, 0)
-	r.Suggestions = make([]string, 0)
-	return wrapResult(r)
-}
-
-func CustomTest2(bundle apimanifests.Bundle) scapiv1alpha3.TestStatus {
-	r := scapiv1alpha3.TestResult{}
-	r.Name = CustomTest2Name
-	r.State = scapiv1alpha3.PassState
-	r.Errors = make([]string, 0)
-	r.Suggestions = make([]string, 0)
-	return wrapResult(r)
-}
-
-func wrapResult(r scapiv1alpha3.TestResult) scapiv1alpha3.TestStatus {
-	return scapiv1alpha3.TestStatus{
-		Results: []scapiv1alpha3.TestResult{r},
-	}
-}
